Update wrappers to f64

+++ bindings/src/data.rs
@@ -7,16 +7,16 @@ use numpy::{PyReadonlyArray1, PyReadonlyArray2};
 use pgbart::math::Matrix;
 use pgbart::pgbart::ExternalData;
 
-type LogpFunc = unsafe extern "C" fn(*const f32, usize) -> std::os::raw::c_float;
+type LogpFunc = unsafe extern "C" fn(*const f64, usize) -> std::os::raw::c_double;
 
 pub struct PythonData {
-    X: Matrix<f32>,
-    y: Vec<f32>,
+    X: Matrix<f64>,
+    y: Vec<f64>,
     logp: LogpFunc,
 }
 
 impl PythonData {
-    pub fn new(X: PyReadonlyArray2<f32>, y: PyReadonlyArray1<f32>, logp: usize) -> Self {
+    pub fn new(X: PyReadonlyArray2<f64>, y: PyReadonlyArray1<f64>, logp: usize) -> Self {
         let X = Matrix::from_vec(X.to_vec().unwrap(), X.shape()[0], X.shape()[1]);
         let y = y.to_vec().unwrap();
 
@@ -27,15 +27,15 @@ impl PythonData {
 }
 
 impl ExternalData for PythonData {
-    fn X(&self) -> &Matrix<f32> {
+    fn X(&self) -> &Matrix<f64> {
         &self.X
     }
 
-    fn y(&self) -> &Vec<f32> {
+    fn y(&self) -> &Vec<f64> {
         &self.y
     }
 
-    fn model_logp(&self, v: &Vec<f32>) -> f32 {
+    fn model_logp(&self, v: &Vec<f64>) -> f64 {
         let logp = self.logp;
         let value = unsafe { logp(v.as_ptr(), v.len()) };
